+++ src/main/java/io/fabric8/maven/docker/util/AnsiLogger.java
@@ -1,5 +1,6 @@
 package io.fabric8.maven.docker.util;
 
+import java.io.*;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
@@ -7,6 +8,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.concurrent.atomic.AtomicInteger;
+import java.util.function.Consumer;
+import java.util.function.Predicate;
 
 import org.apache.maven.plugin.logging.Log;
 import org.codehaus.plexus.util.StringUtils;
@@ -22,7 +25,7 @@ import static org.fusesource.jansi.Ansi.ansi;
  * @author roland
  * @since 31.03.14
  */
-public class AnsiLogger implements Logger {
+public class AnsiLogger implements Logger, Closeable {
 
     // prefix used for console output
     public static final String DEFAULT_LOG_PREFIX = "DOCKER> ";
@@ -31,9 +34,11 @@ public class AnsiLogger implements Logger {
     private final Log log;
     private final String prefix;
     private final boolean batchMode;
+    private final File outputFile;
 
     private boolean isVerbose = false;
     private List<LogVerboseCategory> verboseModes = null;
+    private PrintWriter pw;
 
     // ANSI escapes for various colors (or empty strings if no coloring is used)
     static Ansi.Color
@@ -59,44 +64,78 @@ public class AnsiLogger implements Logger {
     }
 
     public AnsiLogger(Log log, boolean useColor, String verbose, boolean batchMode) {
-        this(log, useColor, verbose, batchMode, DEFAULT_LOG_PREFIX);
+        this(log, useColor, verbose, batchMode, DEFAULT_LOG_PREFIX, null);
     }
 
     public AnsiLogger(Log log, boolean useColor, String verbose, boolean batchMode, String prefix) {
+        this(log, useColor, verbose, batchMode, prefix, null);
+    }
+
+    public AnsiLogger(Log log, boolean useColor, String verbose, boolean batchMode, String prefix, File outpufFile) {
         this.log = log;
         this.prefix = prefix;
-        this.batchMode = batchMode;
+        this.outputFile = outpufFile;
+        if (this.outputFile == null) {
+            this.batchMode = batchMode;
+        } else {
+            this.batchMode = true;
+        }
         checkVerboseLoggingEnabled(verbose);
         initializeColor(useColor);
+        try {
+            initializePrintWriter();
+        } catch (FileNotFoundException e) {
+            e.printStackTrace();
+        }
     }
 
     /** {@inheritDoc} */
     public void debug(String message, Object ... params) {
         if (isDebugEnabled()) {
-            log.debug(prefix + format(message, params));
+            logOrPrintToFile(
+                    log -> true,
+                    log -> log.debug(prefix + format(message, params)),
+                    message,
+                    params);
         }
     }
 
     /** {@inheritDoc} */
     public void info(String message, Object ... params) {
-        log.info(colored(message, COLOR_INFO, true, params));
+        logOrPrintToFile(
+                log -> log.isInfoEnabled(),
+                log -> log.info(colored(message, COLOR_INFO, true, params)),
+                message,
+                params);
     }
 
     /** {@inheritDoc} */
     public void verbose(LogVerboseCategory logVerboseCategory, String message, Object ... params) {
         if (isVerbose && verboseModes != null && verboseModes.contains(logVerboseCategory)) {
-            log.info(ansi().fgBright(BLACK).a(prefix).a(format(message, params)).reset().toString());
+            logOrPrintToFile(
+                    log -> true,
+                    log -> log.info(ansi().fgBright(BLACK).a(prefix).a(format(message, params)).reset().toString()),
+                    message,
+                    params);
         }
     }
 
     /** {@inheritDoc} */
-    public void warn(String format, Object ... params) {
-        log.warn(colored(format, COLOR_WARNING, true, params));
+    public void warn(String message, Object ... params) {
+        logOrPrintToFile(
+                log -> log.isWarnEnabled(),
+                log -> log.warn(colored(message, COLOR_WARNING, true, params)),
+                message,
+                params);
     }
 
     /** {@inheritDoc} */
     public void error(String message, Object ... params) {
-        log.error(colored(message, COLOR_ERROR, true, params));
+        logOrPrintToFile(
+                log -> log.isErrorEnabled(),
+                log -> log.error(colored(message, COLOR_ERROR, true, params)),
+                message,
+                params);
     }
 
     @Override
@@ -211,6 +250,12 @@ public class AnsiLogger implements Logger {
         }
         else {
             Ansi.setEnabled(false);
+        }
+    }
+
+    private void initializePrintWriter() throws FileNotFoundException {
+        if (outputFile != null) {
+            this.pw = new PrintWriter(outputFile);
         }
     }
 
@@ -342,5 +387,21 @@ public class AnsiLogger implements Logger {
             }
         }
         return ret;
+    }
+
+    private void logOrPrintToFile(Predicate<Log> logPredicate, Consumer<Log> logConsumer, String message, Object ... params) {
+        if (outputFile != null && logPredicate.test(log)) {
+            pw.println(format(message, params));
+        } else {
+            logConsumer.accept(log);
+        }
+    }
+
+    @Override
+    public void close() throws IOException {
+        if (pw != null) {
+            pw.flush();
+            pw.close();
+        }
     }
 }